+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ### Version 1.19 - 1.6.2
  - move config to json format
+ - add command `/improvedvanilla`
+   - `/improvedvanilla config show` - provides a clickable link, to find the config file easily (works only in singleplayer)
+   - `/improvedvanilla config reload` - reloads the config file from the file system
+   - `/improvedvanilla config reset` - resets the config to the default values
 
 ### Version 1.19 - 1.6.1
  - fix `Mod Loading has failed` error